Etain's rough work for the first 20%
20% - Describe: 
Good summary of the dataset, repository well laid-out and organised. Reasonable commits to the repository.
Create a git repository and make it available online for the lecturer to clone. The repository should contain all your work for this assessment. Within the repository, create a jupyter notebook that uses descriptive statistics and plots to describe the Boston House Prices dataset. This part is worth 20% of your overall mark.