Etain's work for the first 20% of this project.
20% 
Create a git repository and make it available online for the lecturer to clone. The repository should contain all your work for this assessment. 
Within the repository, create a jupyter notebook that uses descriptive statistics and plots to describe the Boston House Prices dataset. This part is worth 20% of your overall mark.

I will tidy up these rough notes and upload this information to my Jupyter Notebook just before I submit the final version for assessment.